requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.27.2
pyyaml==6.0.2
